@@ -20,8 +20,8 @@
   },
   "homepage": "https://github.com/spring-guides/tut-react-and-spring-data-rest",
   "dependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
+    "react": "^18.3.1",
+    "react-dom": "^18.3.1",
     "rest": "^2.0.0",
     "sockjs-client": "^1.6.1",
     "stompjs": "^2.3.3"
@@ -30,12 +30,11 @@
     "watch": "webpack --watch -d --output ./target/classes/static/built/bundle.js"
   },
   "devDependencies": {
-    "@babel/core": "^7.24.0",
-    "@babel/preset-env": "^7.24.0",
-    "@babel/preset-react": "^7.23.3",
+    "@babel/core": "^7.25.2",
+    "@babel/preset-env": "^7.25.4",
+    "@babel/preset-react": "^7.24.7",
     "babel-loader": "^9.1.3",
-    "webpack": "^5.90.3",
-    "webpack-cli": "^5.1.4",
-    "when": "^3.7.8"
+    "webpack": "^5.94.0",
+    "webpack-cli": "^5.1.4"
   }
 }
