@@ -1,466 +1,247 @@
 'use strict';
 
-const React = require('react');
-const ReactDOM = require('react-dom');
-const ReactDOMClient = require('react-dom/client');
-const when = require('when');
-const client = require('./client');
-
-const follow = require('./follow'); // function to hop multiple links by "rel"
-
-const stompClient = require('./websocket-listener');
+import React, { useState, useEffect, useCallback, useRef } from 'react';
+import ReactDOMClient from 'react-dom/client';
+//import when from 'when';
+import client from './client';
+import follow from './follow'; // function to hop multiple links by "rel"
+import CreateDialog from './components/CreateDialog';
+import EmployeeList from './components/EmployeeList';
+import stompClient from './websocket-listener';
 
 const root = '/api';
 
-class App extends React.Component {
+const App = ({ loggedInManager }) => {
+  const [employees, setEmployees] = useState([]);
+  const [attributes, setAttributes] = useState([]);
+  const [page, setPage] = useState(1);
+  const [pageSize, setPageSize] = useState(2);
+  const [links, setLinks] = useState({});
+  const [schema, setSchema] = useState(null);
+  
+  const stompClientRef = useRef(null);
+  const pageRef = useRef(page);
+  const pageSizeRef = useRef(pageSize);
 
-	constructor(props) {
-		super(props);
-		this.state = {employees: [], attributes: [], page: 1, pageSize: 2, links: {}, loggedInManager: this.props.loggedInManager};
-		this.updatePageSize = this.updatePageSize.bind(this);
-		this.onCreate = this.onCreate.bind(this);
-		this.onUpdate = this.onUpdate.bind(this);
-		this.onDelete = this.onDelete.bind(this);
-		this.onNavigate = this.onNavigate.bind(this);
-		this.refreshCurrentPage = this.refreshCurrentPage.bind(this);
-		this.refreshAndGoToLastPage = this.refreshAndGoToLastPage.bind(this);
-	}
+  useEffect(() => {
+    pageRef.current = page;
+  }, [page]);
 
-	loadFromServer(pageSize) {
-		follow(client, root, [
-			{rel: 'employees', params: {size: pageSize}}]
-		).then(employeeCollection => {
-			return client({
-				method: 'GET',
-				path: employeeCollection.entity._links.profile.href,
-				headers: {'Accept': 'application/schema+json'}
-			}).then(schema => {
-				/**
-				 * Filter unneeded JSON Schema properties, like uri references and
-				 * subtypes ($ref).
-				 */
-				Object.keys(schema.entity.properties).forEach(function (property) {
-					if (schema.entity.properties[property].hasOwnProperty('format') &&
-						schema.entity.properties[property].format === 'uri') {
-						delete schema.entity.properties[property];
-					}
-					else if (schema.entity.properties[property].hasOwnProperty('$ref')) {
-						delete schema.entity.properties[property];
-					}
-				});
+  useEffect(() => {
+    pageSizeRef.current = pageSize;
+  }, [pageSize]);
 
-				this.schema = schema.entity;
-				this.links = employeeCollection.entity._links;
-				return employeeCollection;
-			});
-		}).then(employeeCollection => {
-			this.page = employeeCollection.entity.page;
-			return employeeCollection.entity._embedded.employees.map(employee =>
-					client({
-						method: 'GET',
-						path: employee._links.self.href
-					})
-			);
-		}).then(employeePromises => {
-			return when.all(employeePromises);
-		}).then(employees => {
-			this.setState({
-				page: this.page,
-				employees: employees,
-				attributes: Object.keys(this.schema.properties),
-				pageSize: pageSize,
-				links: this.links
-			});
+  const loadFromServer = useCallback(async (pageSize) => {
+    try {
+      // Fetch employee collection
+      const employeeCollection = await follow(client, root, [{ rel: 'employees', params: { size: pageSize } }]);
 
-		});
-	}
+      // Fetch schema
+      const schemaResponse = await client({
+        method: 'GET',
+        path: employeeCollection.entity._links.profile.href,
+        headers: { 'Accept': 'application/schema+json' }
+      });
 
-	onCreate(newEmployee) {
-		follow(client, root, ['employees']).then(response => {
-			client({
+      // Filter unneeded JSON Schema properties
+      Object.keys(schemaResponse.entity.properties).forEach(property => {
+        if (schemaResponse.entity.properties[property].hasOwnProperty('format') && schemaResponse.entity.properties[property].format === 'uri') {
+          delete schemaResponse.entity.properties[property];
+        } else if (schemaResponse.entity.properties[property].hasOwnProperty('$ref')) {
+          delete schemaResponse.entity.properties[property];
+        }
+      });
 
-				method: 'POST',
-				path: response.entity._links.self.href,
-				entity: newEmployee,
-				headers: {'Content-Type': 'application/json'}
-			})
-		})
-	}
-	
-	onUpdate(employee, updatedEmployee) {
-		if(employee.entity.manager.name === this.state.loggedInManager) {
-			updatedEmployee["manager"] = employee.entity.manager;
-			client({
-				method: 'PUT',
-				path: employee.entity._links.self.href,
-				entity: updatedEmployee,
-				headers: {
-					'Content-Type': 'application/json',
-					'If-Match': employee.headers.Etag
-				}
-			}).then(response => {
-			}, response => {
-				if (response.status.code === 403) {
-					alert('ACCESS DENIED: You are not authorized to update ' +
-						employee.entity._links.self.href);
-				}
-				if (response.status.code === 412) {
-					alert('DENIED: Unable to update ' + employee.entity._links.self.href + '. Your copy is stale.');
-				}
-			});
-		} else {
-			alert("You are not authorized to update");
-		}
-	}
+      setSchema(schemaResponse.entity);
 
-	onDelete(employee) {
-		client({method: 'DELETE', path: employee.entity._links.self.href}
-		).then(response => {},
-			response => {
-				if (response.status.code === 403) {
-					alert('ACCESS DENIED: You are not authorized to delete ' +
-						employee.entity._links.self.href);
-				}
-			}
-		);
-	}
+      setLinks(employeeCollection.entity._links);
 
-	onNavigate(navUri) {
-		client({
-			method: 'GET',
-			path: navUri
-		}).then(employeeCollection => {
-			this.links = employeeCollection.entity._links;
-			this.page = employeeCollection.entity.page;
+      // Set page
+      setPage(employeeCollection.entity.page);
 
-			return employeeCollection.entity._embedded.employees.map(employee =>
-					client({
-						method: 'GET',
-						path: employee._links.self.href
-					})
-			);
-		}).then(employeePromises => {
-			return when.all(employeePromises);
-		}).then(employees => {
-			this.setState({
-				page: this.page,
-				employees: employees,
-				attributes: Object.keys(this.schema.properties),
-				pageSize: this.state.pageSize,
-				links: this.links
-			});
-		});
-	}
+      // Fetch employees
+      const employeePromises = employeeCollection.entity._embedded.employees.map(employee =>
+        client({ method: 'GET', path: employee._links.self.href })
+      );
 
-	updatePageSize(pageSize) {
-		if (pageSize !== this.state.pageSize) {
-			this.loadFromServer(pageSize);
-		}
-	}
+      // Wait for all employee requests to resolve
+      const employees = await Promise.all(employeePromises);
 
-	refreshAndGoToLastPage(message) {
-		follow(client, root, [{
-			rel: 'employees',
-			params: {size: this.state.pageSize}
-		}]).then(response => {
-			if (response.entity._links.last !== undefined) {
-				this.onNavigate(response.entity._links.last.href);
-			} else {
-				this.onNavigate(response.entity._links.self.href);
-			}
-		})
-	}
+      setEmployees(employees);
 
-	refreshCurrentPage(message) {
-		follow(client, root, [{
-			rel: 'employees',
-			params: {
-				size: this.state.pageSize,
-				page: this.state.page.number
-			}
-		}]).then(employeeCollection => {
-			this.links = employeeCollection.entity._links;
-			this.page = employeeCollection.entity.page;
+      if ((schemaResponse) && (schemaResponse.entity) && (schemaResponse.entity.properties)) {
+        setAttributes(Object.keys(schemaResponse.entity.properties));
+      }
+      setPageSize(pageSize);
+    } catch (error) {
+      console.error("Failed to load from server: ", error);
+    }
+  }, [links, schema]);
 
-			return employeeCollection.entity._embedded.employees.map(employee => {
-				return client({
-					method: 'GET',
-					path: employee._links.self.href
-				})
-			});
-		}).then(employeePromises => {
-			return when.all(employeePromises);
-		}).then(employees => {
-			this.setState({
-				page: this.page,
-				employees: employees,
-				attributes: Object.keys(this.schema.properties),
-				pageSize: this.state.pageSize,
-				links: this.links
-			});
-		});
-	}
+  const onCreate = useCallback(async (newEmployee) => {
+    try {
+      // Follow the client to the 'employees' link
+      const response = await follow(client, root, ['employees']);
+      
+      // Make a POST request to create a new employee
+      await client({
+        method: 'POST',
+        path: response.entity._links.self.href,
+        entity: newEmployee,
+        headers: { 'Content-Type': 'application/json' }
+      });
+      
+    } catch (error) {
+      console.error('Error creating employee: ', error);
+    }
+  }, []);
 
-	componentDidMount() {
-		this.loadFromServer(this.state.pageSize);
-		stompClient.register([
-			{route: '/topic/newEmployee', callback: this.refreshAndGoToLastPage},
-			{route: '/topic/updateEmployee', callback: this.refreshCurrentPage},
-			{route: '/topic/deleteEmployee', callback: this.refreshCurrentPage}
-		]);
-	}
+  const onUpdate = useCallback(async (employee, updatedEmployee) => {
+    if (employee.entity.manager.name === loggedInManager) {
+      updatedEmployee["manager"] = employee.entity.manager;
 
-	render() {
-		return (
-			<div>
-				<CreateDialog attributes={this.state.attributes} onCreate={this.onCreate}/>
-				<EmployeeList page={this.state.page}
-							  employees={this.state.employees}
-							  links={this.state.links}
-							  pageSize={this.state.pageSize}
-							  attributes={this.state.attributes}
-							  onNavigate={this.onNavigate}
-							  onUpdate={this.onUpdate}
-							  onDelete={this.onDelete}
-							  updatePageSize={this.updatePageSize}
-  							  loggedInManager={this.state.loggedInManager}/>
-			</div>
-		)
-	}
-}
+      try {
+        const response = await client({
+          method: 'PUT',
+          path: employee.entity._links.self.href,
+          entity: updatedEmployee,
+          headers: {
+            'Content-Type': 'application/json',
+            'If-Match': employee.headers.Etag
+          }
+        });
 
-class CreateDialog extends React.Component {
+      } catch (error) {
+        if (error.status.code === 403) {
+          alert('ACCESS DENIED: You are not authorized to update ' + employee.entity._links.self.href);
+        } else if (error.status.code === 412) {
+          alert('DENIED: Unable to update ' + employee.entity._links.self.href + '. Your copy is stale.');
+        }
+      }
+    } else {
+      alert("You are not authorized to update");
+    }
+  }, [loggedInManager]);
+  
+  const onDelete = useCallback(async (employee) => {
+    try {
+      const response = await client({ method: 'DELETE', path: employee.entity._links.self.href });
+    } catch (error) {
+      if (error.status.code === 403) {
+        alert('ACCESS DENIED: You are not authorized to delete ' + employee.entity._links.self.href);
+      }
+    }
+  }, []);
+  
+  const onNavigate = useCallback(async (navUri) => {
+    try {
+      // Fetch the employee collection
+      const employeeCollection = await client({ method: 'GET', path: navUri });
 
-	constructor(props) {
-		super(props);
-		this.handleSubmit = this.handleSubmit.bind(this);
-	}
+      // Update the links and page state
+      setLinks(employeeCollection.entity._links);
+      setPage(employeeCollection.entity.page);
 
-	handleSubmit(e) {
-		e.preventDefault();
-		const newEmployee = {};
-		this.props.attributes.forEach(attribute => {
-			newEmployee[attribute] = ReactDOM.findDOMNode(this.refs[attribute]).value.trim();
-		});
-		this.props.onCreate(newEmployee);
-		this.props.attributes.forEach(attribute => {
-			ReactDOM.findDOMNode(this.refs[attribute]).value = ''; // clear out the dialog's inputs
-		});
-		window.location = "#";
-	}
+      // Fetch individual employee data
+      const employeePromises = employeeCollection.entity._embedded.employees.map(employee =>
+        client({ method: 'GET', path: employee._links.self.href })
+      );
 
-	render() {
-		const inputs = this.props.attributes.map(attribute =>
-			<p key={attribute}>
-				<input type="text" placeholder={attribute} ref={attribute} className="field"/>
-			</p>
-		);
+      // Wait for all employee requests to resolve
+      const employeesArg = await Promise.all(employeePromises);
 
-		return (
-			<div>
-				<a href="#createEmployee">Create</a>
+      // Update the employees state
+      setEmployees(employeesArg);
+    } catch (error) {
+      console.error("onNavigate error: ", error);
+    }
+  }, []);
 
-				<div id="createEmployee" className="modalDialog">
-					<div>
-						<a href="#" title="Close" className="close">X</a>
+  const updatePageSize = useCallback((newPageSize) => {
+    if (newPageSize !== pageSize) {
+      loadFromServer(newPageSize);
+    }
+  }, [loadFromServer, pageSize]);
 
-						<h2>Create new employee</h2>
+  const refreshAndGoToLastPage = useCallback(async () => {
+    try {
+      // Follow the client to get the employee collection
+      const response = await follow(client, root, [{ rel: 'employees', params: { size: pageSize } }]);
+      
+      // Navigate to the last page if it exists, otherwise to the current page
+      if (response.entity._links.last !== undefined) {
+        await onNavigate(response.entity._links.last.href);
+      } else {
+        await onNavigate(response.entity._links.self.href);
+      }
+    } catch (error) {
+      console.error("Error in refreshAndGoToLastPage: ", error);
+    }
+  }, []);
 
-						<form>
-							{inputs}
-							<button onClick={this.handleSubmit}>Create</button>
-						</form>
-					</div>
-				</div>
-			</div>
-		)
-	}
-}
+  const refreshCurrentPage = useCallback(async () => {
+    // Access current values from refs
+    const currentPage = pageRef.current;
+    const currentPageSize = pageSizeRef.current;
 
-class UpdateDialog extends React.Component {
+    try {
+      // Follow the client to get the employee collection
+      const employeeCollection = await follow(client, root, [{ rel: 'employees', params: { size: currentPageSize, page: currentPage.number } }]);
 
-	constructor(props) {
-		super(props);
-		this.handleSubmit = this.handleSubmit.bind(this);
-	}
+      // Update state with links and page information
+      setLinks(employeeCollection.entity._links);
+      setPage(employeeCollection.entity.page);
 
-	handleSubmit(e) {
-		e.preventDefault();
-		const updatedEmployee = {};
-		this.props.attributes.forEach(attribute => {
-			updatedEmployee[attribute] = ReactDOM.findDOMNode(this.refs[attribute]).value.trim();
-		});
-		this.props.onUpdate(this.props.employee, updatedEmployee);
-		window.location = "#";
-	}
+      // Fetch individual employee details
+      const employeePromises = employeeCollection.entity._embedded.employees.map(employee =>
+        client({ method: 'GET', path: employee._links.self.href })
+      );
 
-	render() {
-		const inputs = this.props.attributes.map(attribute =>
-			<p key={this.props.employee.entity[attribute]}>
-				<input type="text" placeholder={attribute}
-					   defaultValue={this.props.employee.entity[attribute]}
-					   ref={attribute} className="field"/>
-			</p>
-		);
+      // Resolve all employee promises
+      const employees = await Promise.all(employeePromises);
 
-		const dialogId = "updateEmployee-" + this.props.employee.entity._links.self.href;
-		
-				const isManagerCorrect = this.props.employee.entity.manager.name == this.props.loggedInManager;
+      // Update state with employee data
+      setEmployees(employees);
+    } catch (error) {
+      console.error("Error in refreshCurrentPage: ", error);
+    }
+  }, [pageRef, pageSizeRef]);
 
-		if (isManagerCorrect === false) {
-			return (
-					<div>
-						<a>Not Your Employee</a>
-					</div>
-				)
-		} else {
-			return (
-				<div>
-					<a href={"#" + dialogId}>Update</a>
-	
-					<div id={dialogId} className="modalDialog">
-						<div>
-							<a href="#" title="Close" className="close">X</a>
-	
-							<h2>Update an employee</h2>
-	
-							<form>
-								{inputs}
-								<button onClick={this.handleSubmit}>Update</button>
-							</form>
-						</div>
-					</div>
-				</div>
-			)
-		}
-	}
+  
+  useEffect(() => {
+    loadFromServer(pageSize);
+    if (stompClientRef.current === null) {
+      const routes = [
+        { route: '/topic/newEmployee', callback: refreshAndGoToLastPage },
+        { route: '/topic/updateEmployee', callback: () => { refreshCurrentPage(); } },
+        { route: '/topic/deleteEmployee', callback: () => { refreshCurrentPage(); } }
+      ];
 
-}
+      stompClient.register(routes);
+      stompClientRef.current = stompClient;
+      stompClientRef.current.routes = routes; // Store the routes in the ref
+    }
+  }, []);
 
-class EmployeeList extends React.Component {
+  return (
+    <div>
+      <CreateDialog attributes={attributes} onCreate={onCreate} />
+      <EmployeeList
+        page={page}
+        employees={employees}
+        links={links}
+        pageSize={pageSize}
+        attributes={attributes}
+        onNavigate={onNavigate}
+        onUpdate={onUpdate}
+        onDelete={onDelete}
+        updatePageSize={updatePageSize}
+        loggedInManager={loggedInManager}
+      />
+    </div>
+  );
+};
 
-	constructor(props) {
-		super(props);
-		this.handleNavFirst = this.handleNavFirst.bind(this);
-		this.handleNavPrev = this.handleNavPrev.bind(this);
-		this.handleNavNext = this.handleNavNext.bind(this);
-		this.handleNavLast = this.handleNavLast.bind(this);
-		this.handleInput = this.handleInput.bind(this);
-	}
-
-	handleInput(e) {
-		e.preventDefault();
-		const pageSize = ReactDOM.findDOMNode(this.refs.pageSize).value;
-		if (/^[0-9]+$/.test(pageSize)) {
-			this.props.updatePageSize(pageSize);
-		} else {
-			ReactDOM.findDOMNode(this.refs.pageSize).value = pageSize.substring(0, pageSize.length - 1);
-		}
-	}
-
-	handleNavFirst(e) {
-		e.preventDefault();
-		this.props.onNavigate(this.props.links.first.href);
-	}
-
-	handleNavPrev(e) {
-		e.preventDefault();
-		this.props.onNavigate(this.props.links.prev.href);
-	}
-
-	handleNavNext(e) {
-		e.preventDefault();
-		this.props.onNavigate(this.props.links.next.href);
-	}
-
-	handleNavLast(e) {
-		e.preventDefault();
-		this.props.onNavigate(this.props.links.last.href);
-	}
-
-	render() {
-		const pageInfo = this.props.page.hasOwnProperty("number") ?
-		<h3>Employees - Page {this.props.page.number + 1} of {this.props.page.totalPages}</h3> : null;
-
-		const employees = this.props.employees.map(employee =>
-			<Employee key={employee.entity._links.self.href}
-					  employee={employee}
-					  attributes={this.props.attributes}
-					  onUpdate={this.props.onUpdate}
-					  onDelete={this.props.onDelete}
-  					  loggedInManager={this.props.loggedInManager}/>
-		);
-
-		const navLinks = [];
-		if ("first" in this.props.links) {
-			navLinks.push(<button key="first" onClick={this.handleNavFirst}>&lt;&lt;</button>);
-		}
-		if ("prev" in this.props.links) {
-			navLinks.push(<button key="prev" onClick={this.handleNavPrev}>&lt;</button>);
-		}
-		if ("next" in this.props.links) {
-			navLinks.push(<button key="next" onClick={this.handleNavNext}>&gt;</button>);
-		}
-		if ("last" in this.props.links) {
-			navLinks.push(<button key="last" onClick={this.handleNavLast}>&gt;&gt;</button>);
-		}
-
-		return (
-			<div>
-				{pageInfo}
-				<input ref="pageSize" defaultValue={this.props.pageSize} onInput={this.handleInput}/>
-				<table>
-					<tbody>
-						<tr>
-							<th>First Name</th>
-							<th>Last Name</th>
-							<th>Description</th>
-							<th>Manager</th>
-							<th></th>
-							<th></th>
-						</tr>
-						{employees}
-					</tbody>
-				</table>
-				<div>
-					{navLinks}
-				</div>
-			</div>
-		)
-	}
-}
-
-class Employee extends React.Component {
-
-	constructor(props) {
-		super(props);
-		this.handleDelete = this.handleDelete.bind(this);
-	}
-
-	handleDelete() {
-		this.props.onDelete(this.props.employee);
-	}
-
-	render() {
-		return (
-			<tr>
-				<td>{this.props.employee.entity.firstName}</td>
-				<td>{this.props.employee.entity.lastName}</td>
-				<td>{this.props.employee.entity.description}</td>
-				<td>{this.props.employee.entity.manager.name}</td>
-				<td>
-					<UpdateDialog employee={this.props.employee}
-								  attributes={this.props.attributes}
-								  onUpdate={this.props.onUpdate}
-								  loggedInManager={this.props.loggedInManager}/>
-				</td>
-				<td>
-					<button onClick={this.handleDelete}>Delete</button>
-				</td>
-			</tr>
-		)
-	}
-}
-
-const rootReactElement = ReactDOMClient.createRoot(document.getElementById('react'));
-rootReactElement.render(<App loggedInManager={document.getElementById('managername').innerHTML } />);
+const rootElement = document.getElementById('react');
+const rootReactElement = ReactDOMClient.createRoot(rootElement);
+rootReactElement.render(<App loggedInManager={document.getElementById('managername').innerHTML} />);
