@@ -14,8 +14,8 @@ registry.register('text/uri-list', require('./api/uriListConverter'));
 registry.register('application/hal+json', require('rest/mime/type/application/hal'));
 
 module.exports = rest
-	.wrap(mime, { registry: registry })
-	.wrap(uriTemplateInterceptor)
-	.wrap(errorCode)
-	.wrap(template)
-	.wrap(defaultRequest, { headers: { 'Accept': 'application/hal+json' }});
+  .wrap(mime, { registry: registry })
+  .wrap(uriTemplateInterceptor)
+  .wrap(errorCode)
+  .wrap(template)
+  .wrap(defaultRequest, { headers: { 'Accept': 'application/hal+json' } });
